refactor: change model to match be
