+++ dislinkt-fe/.env
@@ -1,1 +1,1 @@
-REACT_APP_API_URL = "http://localhost:8080"
+REACT_APP_API_URL = "https://localhost:7132/api"
