+++ dislinkt-fe/src/components/molecules/SignInForm/SignInForm.tsx
@@ -1,10 +1,9 @@
 import { Field, Formik } from "formik";
-import { useDispatch, useSelector } from "react-redux";
+import { useDispatch } from "react-redux";
 import { useNavigate } from "react-router-dom";
 
-import SignInFormValues from "../../../models/SignInFormValues";
+import SignInFormValues from "../../../models/forms/SignInFormValues";
 import { signIn } from "../../../store/slices/auth";
-import { RootState } from "../../../store/store";
 import signInValidationSchema from "../../../validations/signInValidationSchema";
 import PrimaryButton from "../../atoms/PrimaryButton/PrimaryButton";
 import PrimaryInputField from "../../atoms/PrimaryInputField/PrimaryInputField";
@@ -12,7 +11,7 @@ import PrimaryInputField from "../../atoms/PrimaryInputField/PrimaryInputField";
 import classes from "./SignInForm.module.css";
 
 const signInFormInitialValues: SignInFormValues = {
-  email: "",
+  username: "",
   password: "",
 };
 
@@ -36,10 +35,10 @@ const SignInForm: React.FC = () => {
           <div className={classes.fields}>
             <Field
               component={PrimaryInputField}
-              text="Email"
-              type="email"
-              name="email"
-              value="email"
+              text="Username"
+              type="text"
+              name="username"
+              value="username"
             />
             <Field
               component={PrimaryInputField}
