+++ dislinkt-fe/src/models/UserModel.ts
@@ -1,0 +1,7 @@
+interface UserModel {
+    id: string;
+    email: string;
+    username: string;
+}
+
+export default UserModel;
