+++ dislinkt-fe/src/models/forms/SignInFormValues.ts
@@ -1,5 +1,5 @@
 interface SignInFormValues {
-  email: string;
+  username: string;
   password: string;
 }
