--- dislinkt-fe/src/models/storage/Token.ts
@@ -1,5 +1,0 @@
-interface Token {
-  authToken: string;
-}
-
-export default Token;
