+++ dislinkt-fe/src/validations/signInValidationSchema.ts
@@ -1,9 +1,7 @@
 import * as Yup from "yup";
 
 const signInValidationSchema = Yup.object().shape({
-  email: Yup.string()
-    .email("* Email is not in the valid format")
-    .required("* Required field"),
+  username: Yup.string().required("* Required field"),
   password: Yup.string().required("* Required field"),
 });
